requirements.txt
Flask==2.3.3
coqui-tts==0.27.0
torch==2.1.2
soundfile
numpy<2
gunicorn
